Added review of users

+++ src/main/java/delivery/stork/model/dto/UserDto.java
@@ -14,6 +14,6 @@ public class UserDto {
     Long id;
     String fullName;
     String email;
-    int averageRate;
+    double averageRate;
     String instagram;
 }
